+++ settings.d/90-permission.php
@@ -4,3 +4,9 @@
 $wgGroupPermissions['*']['createaccount'] = false;
 $wgGroupPermissions['*']['edit'] = false;
 $wgGroupPermissions['*']['read'] = false;
+
+// Allow requests from parsoid and ocg
+if ( $_SERVER['REMOTE_ADDR'] === "172.27.1.3" || $_SERVER['REMOTE_ADDR'] === "172.27.1.4" ) {
+    $wgGroupPermissions['*']['read'] = true;
+    $wgGroupPermissions['*']['edit'] = true;
+}